ToolsUI de-clutter. Peel off help menu to separate class.
