+++ ui/src/main/java/ucar/nc2/ui/ToolsModesMenu.java
@@ -30,9 +30,7 @@ class ToolsModesMenu extends JMenu {
 /**
  *
  */
-    ToolsModesMenu(
-        final ToolsUI tui)
-    {
+    ToolsModesMenu(final ToolsUI tui) {
         super("Modes");
         setMnemonic('M');
